Create add-rows-and-attachments.py
This code shows you how you can add rows and attachments into Smartsheet. This works if there are multiple people submitting requests at the same time. It uses the values that were uploaded to the Smartsheet and gets the largest RowID that the user has submitted with the values added in the new row, and adds the attachment. As of right now, this is the only way I can think of to insert rows and attachments (almost) at the same time.